requirements.txt
pvleopard==2.0.4
pvrecorder==1.2.4
tabulate==0.8.10
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.35
langchain-text-splitters==0.0.1
langsmith==0.1.36
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab